requirements.txt
flask
python-dotenv
groq
cohere
googlesearch-python
duckduckgo-search
bs4
pillow
rich
requests
mtranslate
PyPDF2
python-docx